Fixed chat completion request
I noticed when attempting to benchmark google/gemma-2-9b-it on vllm that the order of the "user" and "system" roles in the message object is relevant, and starting with "system" causes the inference server to reject the request with HTTP 400.

This small patch simply reverses the order, which should generally make sense. Also, tested my patch with Llama 3.1 model on vllm.

Signed-off-by: Omar Awile <omar.awile@oracle.com>